A very safe cast for any ^2 based cpu.  Are we expecting any trinary CPUs   any time soon for Apache support?  Cleans up an emit on 32 bit builds.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@94409 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/experimental/cache_util.c
@@ -243,7 +243,7 @@ CACHE_DECLARE(void) ap_cache_msec2hex(apr_time_t j, char *y)
     int i, ch;
 
     for (i = (sizeof(j) * 2)-1; i >= 0; i--) {
-        ch = j & 0xF;
+        ch = (int)j & 0xF;
         j >>= 4;
         if (ch >= 10)
             y[i] = ch + ('A' - 10);
